Performance improvement
If a filter is not specified, a KeyCursor (instead of a regular one) can be called, which significantly improves performance when removing many records at once. Because of this change, the Promise no longer returns the entire record, but only the affected records' keys (primaryKey and, if specified, the value of the index) are returned.